Start redoing hooking code
Currently only hooks replacing compelete functions were redone.
Now hook-specific assembly instructions are generated runtime as a separate
short wrapper, instead of using inline assembly macros at start of hook
functions.

The eventual goal is to have any exe-specific definitions in a few files
that allow easily supporting multiple versions. Getting rid of REG_ABC
macros achieves most of that, and additionally should allow enabling
some compiler features, such as exceptions, compiler-supported profiling,
stack overflow checks etc.

Also removed a little bit of unused code and improved const-correctness
a bit.

+++ src/order.cpp
@@ -9,7 +9,7 @@ Order::Order()
     allocated.Add(first_allocated_order);
 }
 
-Order *Order::Allocate(int order_id, uint32_t position_xy, Unit *target, int fow_unit_id)
+Order *Order::Allocate(uint8_t order_id, uint32_t position_xy, Unit *target, uint16_t fow_unit_id)
 {
     Order *order;
     order = new Order;
